feat: [prd-backend-package-restructure-008] - [Re-home helper-style top-level packages with unclear placement]
