+++ pkg/factory/runtime/buffers/typed_buffer.go
