+++ pkg/service/ingest/filewatcher_multichannel_test.go
@@ -1,4 +1,4 @@
-package listeners
+package ingest
 
 import (
 	"context"
