+++ pkg/service/ingest/filewatcher_test.go
@@ -1,4 +1,4 @@
-package listeners
+package ingest
 
 import (
 	"context"
